+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
 ## [1.0.8] - 2019-07-12
 ### Added
 - `className` prop to all components
 - `coverage` shield to readme
+-  Added `loader` component
 
 ### Changed
 - Changed node-sass from dev-dependencies to dependencies